[2 ~ 4 단계 방탈출 예약 결제 / 배포] 산초(나영서) 미션 제출합니다. (#108)
* test: 예약 정상 추가 테스트 보완

갯수 검증 추가

* feat: 테마에 'price' 속성 추가

* feat: Payment 엔티티 클래스 추가

* feat: 결제 내역 저장 로직 구현

* feat: 이미 결제된 예약은 관리자가 취소할 수 없다는 제약조건 구현

* refactor: 결제 생성 시간 기준 변경

토스에서 받아오는 승인 시간이 아니라, 우리 어플리케이션을 기준으로 DB에 저장되는 시간을 createdAt으로 설정한다

* feat: 응답을 지정한 형식으로 받아오지 못하는 경우 예외 처리

* refactor: 결제 관련 로직 분리

* feat: 결제 승인 api 예외 처리 기능 구현

* refactor: clientHttpRequestFactory 변경

SimpleClientHttpRequestFactory -> JdkClientHttpRequestFactory

* feat: 클라이언트 코드 변경

- 결제 금액과 시간을 볼 수 있도록 변경

* feat: 내 예약 조회 api 응답 형식 변경

- 결제 금액과 시간을 볼 수 있도록 변경

* test: 깨지는 테스트 코드 해결

* chore: 초기 데이터셋 수정

* refactor: 클라이언트 코드 수정

예약 시간과 금액 랜딩 위치 변경

* feat: 스웨거 설정 추가

* feat: admin 컨트롤러 스웨거 데코레이션 추가

* feat: auth 컨트롤러 스웨거 데코레이션 추가

* feat: member 컨트롤러 스웨거 데코레이션 추가

* feat: reservation 컨트롤러 스웨거 데코레이션 추가

* feat: reservationTime 컨트롤러 스웨거 데코레이션 추가

* feat: theme 컨트롤러 스웨거 데코레이션 추가

* feat: waiting 컨트롤러 스웨거 데코레이션 추가

* style: 줄바꿈 및 주석 추가

* refactor: 스웨거 토큰 요청 부분 수정

- Authorize 버튼으로 한번에 인증하고 테스트할 수 있도록

* refactor: 로그 레벨 변경

- 결제 관련 예외가 아니라면 ERROR 가 아니라 WARN 레벨을 사용하도록

* test: 함수 분리

* style: 코드 재정렬

* refactor: 대기 테이블 구조 변경

- reservation 이 member 를 포함하고 있으므로, member 를 따로 저장하지 않고 reservation 에서 가져오도록

* feat: 아이디로 테마 조회하는 기능 추가

* refactor: 테마의 실제 가격으로 결제하도록 클라이언트 코드 수정

* refactor: 결제 처리의 원자성 보장을 위한 함수 수정

- 우선 결제 정보 저장을 하고, 외부 api를 호출하도록 변경한다. 외부 api 호출 시 오류가 발생하면, 결제 정보를 롤백해서 원자성을 유지할 수 있다.

* refactor: H2 를 파일 기반이 아니라 메모리 기반으로 변경

- 파일 기반은 IO 작업이 필요한 반면, 인메모리는 훨씬 빠르게 접근할 수 있다.

* refactor: 예약 정보 조회 함수 수정

- 예약 정보와 결제 정보를 하나의 쿼리로 가져올 수 있도록 변경
- ReservationRepository 에 JPQL을 통해 직접 dto 를 반환하도록 하는 함수 생성

* refactor: 예약 정보 조회 함수 수정

- 예약 정보와 결제 정보를 하나의 쿼리로 가져올 수 있도록 변경
- ReservationRepository 에 JPQL을 통해 직접 dto 를 반환하도록 하는 함수 생성

* test: 테스트 데이터 수정으로 깨지는 테스트 해결

* refactor: 결제가 안된 예약에 대한 처리 추가

- 결제가 되지 않아 결제 시각이 없는 경우 '결제 대기 중'을 보여주도록

* refactor: 연속된 getter 함수 호출 개선

* refactor: 예약과 관련된 API 허용 범위 수정

- 전체 예약 조회는 관리자만 가능
- 예약 취소는 생성한 사람만 가능

* refactor: 테마와 관련된 API 허용 범위 수정

- 테마 생성, 삭제는 관리자만 가능

* refactor: 예약 시간과 관련된 API 허용 범위 수정

- 예약 시간 생성, 삭제는 관리자만 가능

* refactor: 회원 조회과 관련된 API 허용 범위 수정

- 관리자만 전체 회원 정보 조회 가능

* refactor: 대기 삭제 관련된 API 허용 범위 수정

- 대기를 생성한 사람만 삭제 가능

* refactor: 파일 경로 수정

* refactor: 파일 경로 수정

* refactor: 에러 코드 수정

- 삭제 권한이 없을 때의 예외 응답 상태 코드를 403으로 수정

+++ src/main/java/roomescape/registration/domain/reservation/service/ReservationService.java
@@ -5,8 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.admin.domain.FilterInfo;
 import roomescape.admin.dto.AdminReservationRequest;
 import roomescape.admin.dto.ReservationFilterRequest;
-import roomescape.client.payment.TossPaymentClient;
-import roomescape.client.payment.dto.PaymentConfirmToTossDto;
 import roomescape.exception.RoomEscapeException;
 import roomescape.exception.model.MemberExceptionCode;
 import roomescape.exception.model.ReservationExceptionCode;
@@ -14,6 +12,8 @@ import roomescape.exception.model.ReservationTimeExceptionCode;
 import roomescape.exception.model.ThemeExceptionCode;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
+import roomescape.payment.repository.PaymentRepository;
+import roomescape.payment.service.PaymentService;
 import roomescape.registration.domain.reservation.domain.Reservation;
 import roomescape.registration.domain.reservation.dto.ReservationRequest;
 import roomescape.registration.domain.reservation.dto.ReservationResponse;
@@ -39,18 +39,20 @@ public class ReservationService {
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
     private final WaitingRepository waitingRepository;
-    private final TossPaymentClient tossPaymentClient;
+    private final PaymentService paymentService;
+    private final PaymentRepository paymentRepository;
 
     public ReservationService(ReservationRepository reservationRepository,
                               ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository,
                               MemberRepository memberRepository, WaitingRepository waitingRepository,
-                              TossPaymentClient tossPaymentClient) {
+                              PaymentService paymentService, PaymentRepository paymentRepository) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.memberRepository = memberRepository;
         this.waitingRepository = waitingRepository;
-        this.tossPaymentClient = tossPaymentClient;
+        this.paymentService = paymentService;
+        this.paymentRepository = paymentRepository;
     }
 
     @Transactional
@@ -66,8 +68,8 @@ public class ReservationService {
         Reservation saveReservation = new Reservation(reservationRequest.date(), time, theme, member);
         Reservation reservation = reservationRepository.save(saveReservation);
 
-        PaymentConfirmToTossDto paymentConfirmToTossDto = PaymentConfirmToTossDto.from(reservationRequest);
-        tossPaymentClient.sendPaymentConfirm(paymentConfirmToTossDto);
+        paymentService.sendConfirmRequestAndSavePayment(reservationRequest, reservation);
+
         return ReservationResponse.from(reservation);
     }
 
@@ -117,17 +119,16 @@ public class ReservationService {
 
     @Transactional(readOnly = true)
     public List<ReservationResponse> findMemberReservations(long id) {
-        List<Reservation> reservations = reservationRepository.findAllByMemberId(id);
-
-        return reservations.stream()
-                .map(ReservationResponse::from)
-                .toList();
+        return reservationRepository.findAllReservationsWithPaymentsByMemberId(id);
     }
 
     @Transactional
-    public void removeReservation(long reservationId) {
-        Optional<Waiting> waiting = waitingRepository.findFirstByReservationIdOrderByCreatedAt(reservationId);
+    public void removeReservation(long reservationId, Long memberId) {
+        if (paymentRepository.existsByReservationId(reservationId)) {
+            throw new RoomEscapeException(ReservationExceptionCode.RESERVATION_ALREADY_PAID_CAN_NOT_DELETE_EXCEPTION);
+        }
 
+        Optional<Waiting> waiting = waitingRepository.findFirstByReservationIdOrderByCreatedAt(reservationId);
         if (waiting.isEmpty()) {
             reservationRepository.deleteById(reservationId);
             return;
@@ -135,7 +136,12 @@ public class ReservationService {
 
         Reservation reservation = reservationRepository.findById(reservationId)
                 .orElseThrow(() -> new RoomEscapeException(ReservationExceptionCode.RESERVATION_NOT_EXIST));
-        reservation.setMember(waiting.get().getMember());
+        if(reservation.getMember() != memberRepository.findMemberById(memberId)
+                .orElseThrow(() -> new RoomEscapeException(MemberExceptionCode.MEMBER_NOT_EXIST_EXCEPTION))){
+            throw new RoomEscapeException(ReservationExceptionCode.ONLY_OWNER_CAN_DELETE);
+        }
+
+        reservation.setMember(waiting.get().getReservation().getMember());
 
         reservationRepository.save(reservation);
         waitingRepository.deleteById(waiting.get().getId());
